+++ africanus/rime/phase.py
@@ -7,14 +7,15 @@ from __future__ import print_function
 from functools import wraps
 import math
 
-import numba
 import numpy as np
 
 from africanus.constants import minus_two_pi_over_c
-from africanus.util.docs import DocstringTemplate, on_rtd
+from africanus.util.docs import DocstringTemplate
+from africanus.util.numba import generated_jit
 from africanus.util.type_inference import infer_complex_dtype
 
 
+@generated_jit(nopython=True, nogil=True, cache=True)
 def phase_delay(lm, uvw, frequency):
     # Bake constants in with the correct type
     one = lm.dtype(1.0)
@@ -53,20 +54,27 @@ def phase_delay(lm, uvw, frequency):
     return _phase_delay_impl
 
 
-if not on_rtd():
-    jitter = numba.generated_jit(nopython=True, nogil=True, cache=True)
-    phase_delay = jitter(phase_delay)
-
-
 PHASE_DELAY_DOCS = DocstringTemplate(
     r"""
     Computes the phase delay (K) term:
 
     .. math::
 
-        & {\Large e^{-2 \pi i (u l + v m + w n)} }
+        & {\Large e^{-2 \pi i (u l + v m + w (n - 1))} }
 
-        & \textrm{where } n = \sqrt{1 - l^2 - m^2} - 1
+        & \textrm{where } n = \sqrt{1 - l^2 - m^2}
+
+    Notes
+    -----
+
+    Corresponds to the complex exponential of the `Van Cittert-Zernike Theorem
+    <https://en.wikipedia.org/wiki/Van_Cittert%E2%80%93Zernike_theorem_>`_.
+
+    `MeqTrees
+    <https://github.com/ska-sa/meqtrees-timba/blob/
+    6a7e873d4d1fe538981dec5851418cbd371b8388/MeqNodes/src/PSVTensor.cc#L314_>`_
+    uses a positive sign convention and so any UVW coordinates must be inverted
+    in order for their phase delay terms (and therefore visibilities) to agree.
 
     Parameters
     ----------