+++ HISTORY.rst
@@ -2,8 +2,20 @@
 History
 =======
 
-0.1.3 (2018-03-28)
+0.1.5 (YYYY-MM-DD)
 ------------------
+* Fix radec sign conversion in wsclean sky model (:pr:`96`)
+* Full Time and Channel Averaging Implementation (:pr:`80`)
+* Support integer seconds in wsclean ra and dec columns (:pr:`91`, :pr:`93`)
+* Fix ratio computation in Gaussian Shape (:pr:`89`, :pr:`90`)
+
+0.1.4 (2019-03-11)
+------------------
+* Support `complete` and `complete-cuda` to support non-GPU installs (:pr:`87`)
+* Gaussian Shape Parameter Implementation (:pr:`82`, :pr:`83`)
+* WSClean Spectral Model (:pr:`81`)
+* Compare predict versus MeqTrees (:pr:`79`)
+* Time and channel averaging (:pr:`75`)
 * cupy implementation of `predict_vis` (:pr:`73`)
 * Introduce transpose in second antenna term of predict (:pr:`72`)
 * cupy implementation of `feed_rotation` (:pr:`67`)